Added Foreman and Procfile to run processes
Use Foreman to run both servers at once, which is a much nicer/cleaner solution.  I don't love introducing a Ruby dependency in a Node system, but I think it's worth it to be able to monitor both server in the same terminal window and start/stop them both at the same time.