Add tests and refactor file reading
Turned the file reading into a utility method that wraps the file reading
actions in a Promise, so that the code is easier to use internally.  This also
makes the code easier to test.  I was able to add test handling to the reader
utility to get around the fact that it's hard to actually do tests with real
file objects.